consumer/requirements.txt
kafka-python
pandas
psycopg2-binary
pydantic
